fixing bugs...

+++ AliceClientApp/main.go
@@ -16,7 +16,7 @@ func main() {
 	flag.StringVar(&txFlag, "t", txFlag, "set transactin type")
 	flag.IntVar(&ChannelVersion, "cv", ChannelVersion, "set channel version")
 	flag.Float64Var(&Coins, "coins", Coins, "set deposit coins")
-	flag.Float64Var(&PreTxId, "pti", PreTxId, "set previous transaction id")
+	flag.Int64Var(&PreTxId, "pti", PreTxId, "set previous transaction id")
 	flag.Parse()
 	err := Execute(txFlag, PreTxId, uint8(ChannelVersion), Coins)
 	if err != nil {
